refactor: Moved worker logic in `UrlArrayLoader` to `SharedWorkerPool` for extensibility
