+++ src/colorizer/loaders/UrlArrayLoader.ts
@@ -1,15 +1,11 @@
 // sort-imports-ignore
 import { DataTexture } from "three";
-import workerpool from "workerpool";
-
-// Vite import directive for worker files! See https://vitejs.dev/guide/features.html#import-with-query-suffixes.
-// @ts-ignore Ignore missing file warning
-import WorkerUrl from "./workers/urlLoadWorker?url&worker";
 
 import { FeatureArrayType, FeatureDataType } from "../types";
 import { infoToDataTexture } from "../utils/texture_utils";
 
 import { ArraySource, IArrayLoader } from "./ILoader";
+import SharedWorkerPool from "../workers/SharedWorkerPool";
 
 export class UrlArraySource<T extends FeatureDataType> implements ArraySource<T> {
   array: FeatureArrayType[T];
@@ -44,18 +40,12 @@ export class UrlArraySource<T extends FeatureDataType> implements ArraySource<T>
 }
 
 export default class UrlArrayLoader implements IArrayLoader {
-  private workerPool: workerpool.Pool;
+  private workerPool: SharedWorkerPool;
+  private disposeOfWorkerPool: boolean;
 
-  constructor() {
-    // TODO: Maintain a single worker pool for all loaders/all asynchronous operations in the app?
-    this.workerPool = workerpool.pool(WorkerUrl, {
-      workerOpts: {
-        // Set worker type to undefined (classic) in production to fix a Vite issue where the application
-        // crashes when loading a module worker.
-        // Copied from https://github.com/josdejong/workerpool/tree/master/examples/vite.
-        type: import.meta.env.PROD ? undefined : "module",
-      },
-    });
+  constructor(workerPool?: SharedWorkerPool) {
+    this.disposeOfWorkerPool = workerPool === undefined;
+    this.workerPool = workerPool ?? new SharedWorkerPool();
   }
 
   /**
@@ -71,15 +61,19 @@ export default class UrlArrayLoader implements IArrayLoader {
    */
   async load<T extends FeatureDataType>(url: string, type: T, min?: number, max?: number): Promise<UrlArraySource<T>> {
     if (!url.endsWith(".json") && !url.endsWith(".parquet")) {
-      throw new Error(`Unsupported file format for URL array loader: ${url}`);
+      throw new Error(
+        `Encountered unsupported file format when loading data. URL must end in '.parquet' or '.json': ${url}`
+      );
     }
-    const { data, textureInfo, min: newMin, max: newMax } = await this.workerPool.exec("load", [url, type]);
+    const { data, textureInfo, min: newMin, max: newMax } = await this.workerPool.load(url, type);
 
     const tex = infoToDataTexture(textureInfo);
     return new UrlArraySource<T>(data, tex, type, min ?? newMin, max ?? newMax);
   }
 
   dispose(): void {
-    this.workerPool.terminate();
+    if (this.disposeOfWorkerPool) {
+      this.workerPool.terminate();
+    }
   }
 }
